import/requirements.txt
astroid==2.4.2
Babel==2.8.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
csvalidate==1.1.1
Flask==1.1.2
Flask-Babel==2.0.0
Flask-CSV==1.2.0
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
Flask-WTF==0.14.3
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-core==1.4.1
google-cloud-storage==1.31.1
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
pathlib==1.0.1
pg8000==1.16.5
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
scramp==1.2.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
